MXS-4529: Go into AUTH state before STARTUP
The authentication happens before the actual startup happens. This way
all the authentication related messages are handled in State::AUTH and
the final startup processing is done in State::STARTUP.

Removed the m_auth_method field since it was not really needed as the
messages always include it and the authentication mechanism can have
multiple values for it. For example the SASL authentication mechanism
uses values 10, 11 and 12 for the authentication message types and thus
storing it is not that useful.