MXS-4529: Store session command history in a class
Since the session command configuration was moved to the MaxScale core
in 6.4, the history itself should be a part of the core. As not all
protocols need it, an opt-in mechanism was needed. Moving the
functionality that was in the MariaDB protocol into a generic History
class makes it reusable for the Postgres protocol.

The comparison of the results is still superficial. The only comparison
that is done is whether the command was successful or not. A more
detailed comparison would be appropriate and could be added by storing a
custom response type in the history.