MXS-4533: Clarify SCRAM authentication
Adding helper functions that return values makes the code much easier to
read. With return values instead of output arrays, the code is very
similar to the abstract version described in the RFC 5802 SCRAM
specification.

Modified the to_base64 and from_base64 functions to accept a wider range
of types. Instead of requiring a vector in the to_base64 overload, using
a template allows anything with a data() and size() methods to be
used. The from_base64 can also use a string_view since it uses a pointer
and a size internally.