MXS-4529: Add History::Subscriber
The act of tracking responses and comparing them to the history can be
moved into the history class as well. The Subscriber class handles the
actual comparisons and the history callback requests.

The m_history_responses list was redundant. When the number of expected
replies drops down to zero, the history replay is complete. This allows
the normal result comparison code to be used to verify the history.