MXS-4544: Rename `changeNodeSize` method to `onNodeResized`
`onNodeResized` indicates that the method is called after changing
the size of the node which is the current usage of the method. On
the other hand, `changeNodeSize` is a wrong name because the method
isn't used to change the size of the node, but to update value
of the `dynNodeSizeMap`.