MXS-4552: Always track requests and their responses
Some core features, such as LOAD DATA LOCAL INFILE and COM_CHANGE_USER
tracking, require the result stream to be at a known point before
MaxScale can safely perform the action. Since the tracking was nearly
always enabled, due to connection_keepalive being on by default, this
caused hard to track bugs in code that was written with the assumption
that request tracking was always done.