MXS-4532: Remove useless code
The debug assertion in mxs::Backend::~Backend prevented the automatic
cleanup from taking place. It didn't really protect anything and only
ended up requiring the same code to be written in the routers that used
it.

Also removed the m_closed variable from schemarouter. It's dead code as
it's never reached. If it were, there would be something seriously wrong
in MaxScale.