MXS-4410 Do not loose that the state is being changed
It is possible that a routing state of CHANGING_[DB|ROLE|USER]
is overwritten by RECORD_HISTORY, leading to the change being
missed.

Now the fact that a change is being made, is tracked separately
so that even if a routing state of CHANGING_STATE is replaced
with RECORD_HISTORY, the state change will not be lost.