MXS-4532: Move response count code to the protocol
Replaced the MariaDB-specific code with a call to the new will_respond()
member function of MXS_SESSION::ProtocolData. This moves the
implementation to the protocol but keeps the API similar for all
modules. This allows the same logic to be used for tracking whether
results are expected or not, regardless of the protocol.