MXS-4532: Add xrouter module
Right now the xrouter module is mostly a skeleton class that is similar
to readconnroute but where readconnroute pipes everything through, the
xrouter tracks the response state. This means that the xrouter is
capable of halting the routing to perform operations that need to be
done on all nodes.

The code that routes to all nodes is not yet active as the query
classifier need to be extended to be able to detect more query types.