Implement WorkerGlobal using std::shared_ptr<T>
This reduces the amount of copying and memory used by it. At the same
time, non-const access to the value is removed to make it clear that the
locally cached value should not be modified and that a read-modify-write
cycle is needed.

The only real usage of the WorkerGlobal where the local value was
modified was when the Service modified the Data struct. This avoided one
copy but with the removal of the non-const methods, this can no longer
be done. Given that the total number of copies still decreases in this
case, it should be an acceptable compromise.

Moved the collect_values() function into WorkerLocal for which it makes
more sense: the class behaves like a thread-local member variable and
the act of collecting the values makes more sense for it than the
globally updatable WorkerGlobal. All uses where the local WorkerGlobal
value was modified and the collect_values() function was used can now be
replaced with the WorkerLocal class.

Removed the legacy workerlocal.cc source file and added the relevant
parts into the header. This amounted to a grand total of three lines of
code so inlining it is acceptable.

The WorkerGlobal that was used in the servers to track the GTID
positions can be replaced with a WorkerLocal version. This also removes
the need to go through the MainWorker as the RoutingWorker::broadcast
call can directly manipulate the local value.

Removed the operator T functions from both WorkerLocal and
WorkerGlobal. They weren't used and it ended up causing situations where
it was possible call assign() with the WorkerGlobal itself.