MXS-4537 Fix undefine connection
This bug occurs after moving `initEntities` function to orm
modules. Connection is not bound to the right query tab because
the action in vuex can only accept 1 parameter. As a result,
the connection isn't bound to the newly inserted query tab.

This also removes the leftover `name` field in `openQueryTabConn`
action.