MXS-4532: Add protocol_data() helper function
As more of the functionality needed to implement filters and routers is
moved into the protocol, accessing it via the m_pSession pointer starts
taking up screen space. This also allows the return value to be made
const which helps avoid unintended modifications.

To be able to forward declare the class, the ProtocolData was moved out
from MXS_SESSION and into the maxscale namespace.