Fix problems found with UBSAN
The act of canceling the dcalls in the destructor of the base Worker
class ended up calling member functions of objects that were already
destroyed. This is what UBSAN was reporting with the error about using
the wrong type of pointer for a derived type. The correct thing to do is
to cancel the dcalls before the object to which the call would be made
is destroyed.

The integer reading functions in the MariaDB protocol module still used
direct addressing instead of memcpy like newer releases do.