MXS-4529: Use std::string_view for variables
Using std::string_view as the return value of get_variables removes the
extra copy that was done. The use of std::string was no longer required
as all the calling code could be converted to use std::string_view.

Using a std::map allows transparent comparators to be used. This removes
the need to construct a value when doing lookups into the
container. This should reduce the amount of wasted work as all lookups
used a string constant as the search key.