MXS-4529: Track Postgres transaction state
The transaction state is conveniently stored in the ReadyForQuery
message. This does not detect the edges of the transaction (i.e. when
it's about to start or stop), for this the SQL itself must be parsed.

Fixed the autocommit state to return true until it is explicitly
tracked. Postgres has autocommit enabled by default.