MXS-4529: Simplify LOAD DATA LOCAL INFILE tracking
The LOAD_DATA_END state was redundant and was only used to track the
number of bytes that were sent. The byte information was only used for a
info level message and moving that message into the protocol greatly
simplifies the query classifier.

Also implemented support for inbound COPY commands. The xrouter does not
yet know how to handle this but with this simplification of the states,
the tracking is very simple to implement and a boolean flag that tells
whether the load is ongoing or not can be used. Eventually, the tracking
should be done by some reusable component.