MXS-4531 Move QC_STMT_INFO to GWBUF::ProtocolInfo
The variable inside GWBUF was made mutable and the member
function GWBUF::set_protocol_info() was made const. That opens
the way for making all Parser functions take a const GWBUF.

"ProtocolInfo" is neutral enough that it does not stick out like
a sore thumb on GWBUFs travelling from the server to the client.
Right now the assumption is that on GWBUFs travelling from the
client to the server, it *is* parser information.