MXS-5432: Lock main node during DDLs
Locking the main node makes sure that DDL execution is serialized when
done through the xrouter. This guarantees that when the client gets the
result, the DDL has been completed on both the main node as well as on
all the secondary nodes. Failures in secondary nodes are still only
handled by closing the corresponding connection and proper fencing of
failed nodes still needs to be implemented.

The addition of locking also requires that a predetermined node is
chosen as the main node. This means that the randomly selected node can
no longer be used for the DDL commands and a new concept of a "solo
node" must be introduced. In practice this means that the solo node is
now the randomly chosen one and the main node is the first declared node
in the cluster. This is not ideal as changes in the order of nodes
during runtime can cause the main node to shift somewhere else.

In theory it would still be possible to use a randomly selected node for
the DDLs but it would require locking all the nodes in a fixed order
instead of locking just the main node. This would be more complex as it
would involve locking multiple nodes as well as more prone to deadlocks
caused by configuration changes. As there's no real practical benefit to
using a random node for DDLs over a fixed one, this approach was not
chosen.

The locking and unlocking SQL is also modifiable. This is done as the
advisory locks in PostgreSQL do not use strings but numbers. The value
in the default SQL was the current Unix timestamp in seconds at the time
when this commit was being done. The value ended up being 1679475768
which is hopefully "random" enough that it won't conflict with existing
locks used by applications.