MXS-4532: Fence out bad nodes with maintenance mode
This prevents failing nodes from being used by both new sessions as well
as existing ones.

Existing sessions will continue using the nodes for any ongoing
operations but once they reach an idle state, unusable nodes are
discarded. If the solo node or the main node gets discarded or cannot be
connected to, the session is closed.

This also causes nodes to be discarded that lose the Running state after
the session has started.