Fix unit tests with ASAN and GCC 10
GCC 10 has a broken ASAN implementation on Debian 11 and it segfaults if
it's used with crypt. The same version appears to be used on SLES 15 as
it fails with similar problems. Unfortunately the --no-as-needed flag
does not prevent SLES 15 from failing. For this reason the tests are
skipped if GCC 10 is used with ASAN.