MXS-4529: Prepare for backlog interruptions
The backlog state was transient and is not really needed. The removal of
this state also simplifies the code and makes it easier to see where
code is accessed without needing to mentally track the state variables.

Any code that halts the packet processing must be aware of the backlog
and where previously a state transition into BACKLOG would've been done,
instead a call send_backlog() needs to now be made after transitioning
into the ROUTING state.