MXS-4529: Move query tracking into sub-functions
This allows the same code to be used to track the responses of session
commands when they are executed as a part of the history.

The old code suffered from a problem where the upload size for queued
responses was not tracked. The problem is fixed now that the same
function is used to start tracking a query.