MXS-4529: Use mxs::History in postgresprotocol
PgBackendConnection now uses mxs::History to handle the execution and
tracking of session commands. The history itself is not yet recorded as
the client protocol is what must implement it.

The implementation is functionally very similar to the one in the
MariaDB protocol. The only notable change is that the sending of the
history is not always done: if no history exists, the state machine goes
directly into the ROUTING state and skips the HISTORY state
completely. This simplifies the code by removing the need to check for
an "empty" history when handling the history responses.