MXS-4531 Add -c flag to compare
With the -c flag it is now possible to specify what the expected
return value is (only for type and operation). So, together with
the flags -x and -p it is now possible to verify that the type or
operation of certain type of statements will be something
particular.

For instance, with the flags

  -p "operation" -x "ALTER TABLE" -c "OP_ALTER_TABLE"

we can check that the operation of all "ALTER TABLE" statements
is reported to be OP_ALTER_TABLE.

In practice, it is now possible to check that the operation of
all "ALTER TABLE" statements in the Postgres test-suite is reported
to be OP_ALTER_TABLE. And similarly for all "CREATE TABLE" etc.
statements.