MXS-4532: Add COPY support to xrouter
The COPY TO command that returns a data stream from the server was
already supported. This commit adds the necessary changes needed to
support the COPY FROM command. The COPY FROM command is similar to the
LOAD DATA LOCAL INFILE command in MariaDB and behaves in a nearly
identical manner.

CopyData and CopyDone must be treated as messages that do not return a
response. They are always done after a Query message and they do not
generate any extra responses. This is similar to how in MariaDB the LOAD
DATA LOCAL INFILE packets do not generate responses and merely amend to
the initial COM_QUERY that starts it.

The xrouter has a separate state where the COPY FROM data uploads are
handled. Although the SOLO state could be reused for this, it would
require treating CopyDone as a command that returns a response when in
reality it doesn't.