MXS-4529: finish_connection() must reset m_subscriber
The lifetime of the shared protocol data is tied to the session lifetime
of the session. Freeing the last DCB releases the last reference to the
session which means that the shared data cannot be accessed in the
destructors.

Added comments about this into the protocol modules where the call is
made as well as into mxs::History::subscribe().