MXS-4544 Add function to draw entity relation link
This gets inspiration from `https://app.diagrams.net/` for the
"entity relation" link in terms of the way the shape of the link
being changed depends on the position of the nodes.

This also adjusts `handleCollision` to add 100px as an extra radius
to prevent nodes and its link from being overlapped.