MXS-4532: Execute node role SQL dynamically
Executing the `main_sql` or `secondary_sql` right before it's needed
removes the need for the INIT state.

It will also simplify the handling of temporary tables. With this
approach the handling of temporary tables can be done by just switching
the role from secondary to main before treating them as normal
single-node commands.