MXS-4531 Move mysql_plus.hh and PacketTracker to Smartrouter
The content of mysql_plus.hh could be moved to maxsql/mariadb.hh.
However, the classes were initially copied from masking/mysql.hh
and after that they have diverged. The differences should be
dealt with after which the functionality can be made available
in maxsql. But right now the main point is to remove MySQL
from include/maxscale.

PacketTracker could also be in maxsql but as it depends on
mysql_plus.hh it is moved to smartrouter for the time being.