MXS-4544 Reposition overlapped points
An attribute of a table can be associated with multiple tables, so the
connected points are overlapped which makes it difficult to differentiate
the links.

This adds `repositionOverlappedPoints` method to reposition points that
are connected to the same position of the entity.

This also simplifies the HTML for entity table.