MXS-4544 Change `getLinkCtr` fuction to use `id`
A node can have two links to another node, and finding the HTML
elements for these links using the source and target node IDs is
reasonable for dragging events. However, this method is not suitable
for redrawing a specific link because it cannot guarantee the
uniqueness of the link.

This switches to use `id` attribute of the link.
This also renames `idPath` parameter to `nodeIdPath` to avoid any
confusion with the link id.