Disable -fsanitize=vptr
There are cases where the same symbols end up being used in multiple
modules. This mostly happens with the mariadbprotocol module but it will
also be a problem in qc_sqlite once it starts using a C++ class in the
API.

Also disabled recovery for UBSAN so that MaxScale aborts as soon as a
problem is found instead of continuing normally.