MXS-4529: Add connection pooling support
The postgres protocol now supports limited connection reuse for users
that come from the same host with the same username. This is not ideal
for good reusability of connections but it does enable it for
installations where the same user and database is often used.

As Postgres does not support proxy protocol, the clients will always
appear to come from the MaxScale host. This also means that the client
IP address does not need to be inspected when a connection is considered
for reuse.

The client protocol now also sets the appropriate flags in the session
to mark that the connections created from it are valid for use in
connection pooling.