MXS-4576: Return a reference from protocol_data()
This commit changes the requirements for the protocol modules. The
protocol modules must now always set the mxs::ProtocolData for each
session. The only protocol that did not set it was the CDC protocol.

With the change in the requirements, the functions in the routers and
filters can now return references to both the protocol data and the
protocol itself.

Also fixed a minor formatting problem in the filter headers. Uncrustify
doesn't accept spaces around curly braces.