MXS-4576: Add MariaDB protocol support to xrouter
The router now supports both the MariaDB and the Postgres
protocols. Right now there's no functional difference as the functions
implemented in the derived classes are not yet used.

There will be a need to have custom behavior for different protocol and
database implementations. Splitting off the functionality into a derived
class that does the database-specific action makes it possible to keep
the common code in the base XRouterSession class.

This also makes it also easier to configure the router. The need to
specify the full SQL is no longer needed and can be determined based on
the protocol that is used. The `main_sql` and `secondary_sql` can be
removed and the `lock_sql` and `unlock_sql` can be replaced with a
`lock_id` field that defines which lock to use.