MXS-4576: Get SCRAM-SHA-256 secrets from a password
The password salting is done using the utility functions found in
libgsasl. In the `salt_password` helper, the results are then converted
into the same format that is used when users are being restored from a
logical backup.

The libgsasl and libidn libraries are available in most repositories in
some form but none of them provide a consistent set of
functionality. For this reason, bundling them into MaxScale as a static
libraries is the most robust approach that guarantees consistent
behavior on all platforms.