MXS-4544 Extends `graph-nodes` to emit mouseover and mouseout events
Mouseover a node will highlight the style of links in the same way that
dragging the node does.

This also adjusts `repositionOverlappedPoints` method to redraw also
invisible links and adds `stroke` property to update the color of
the link and its markers. The actual color for the `stroke` property
will be determined in a later step in the `MxsErd` component because
the links may have different colors.