MXS-4544 Add `GraphConfig` class and take it into uses
This class acts as a global config that shares link, marker and
linkShape config to `EntityLink, EntityMarker, EntityLinkShape` classes.
Updating config would be a lot easier.

This also tweaks the `EntityMarker` and its cardinality symbols to make
sure the links are drawn to the edge of the entity, the markers will
place above the links.