MXS-4577 Use default_helper() in compare.
Now possible to compare the behaviour of a MariaDB parser plugin
to that of a Postgres parser plugin.

Problem is that SQL in a Postgres packet has a trailing NULL,
while SQL in a MariaDB packet does not. This is significant,
because the Postgres parser assumes NULL-terminated SQL, while
the MariaDB parser does not.

Thus, it is straightforward to feed a MariaDB parser with SQL
from a Postgres packet, but not the other way around.