Add mxb::sv_strcasestr
Searching for tokens in std::string_views is tricky without a helper
function like this.

Moved the implementation of the one in the xrouter module to
maxbase. The function signature needed changing as iterators aren't that
useful when searching with string literals. The return value is also
better expressed as an offset into the search string: this allows it to
be used as the argument to std::string_view::substr for further searches
after the original match.

Added a unit test that makes sure it behaves like strcasestr does when
used with string constants.