Use milliseconds in delay_routing()
Makes it clear what units are being used at the call site. Since the
units that were used internally were already milliseconds, there's no
real point in forcing callers to use seconds.

This is also be needed by the throttlefilter that uses milliseconds as
the unit. A follow-up commit will replace the use of dcall with
delay_routing which also solves a possible bug in it.