MXS-4544 Refactor the style update for `Link, EntityMarker`
Leverages the `update` lifecycle hook of `d3.js`, the styles
for certain links and entity markers can be updated by simply
changing the `linkStyles` object of the link.

To change the styles temporarily when hovering over a node or a link,
`tmpUpdateLinksStyle` method is added to the Link class.

Also, mxs-dag-graph has been adjusted to allow mutability of
the links data as d3-dag links return immutable data.

The disadvantage of this approach is that links must be redrawn after
the link data has been mutated. However, D3 is smart enough to
redraw only the updated links, rather than all of them.

This approach allows to have different styles for certain links
ER diagram will use this feature to temporarily change the color
of FK links and PK links.