MXS-4585: Fix debug assertion on shutdown
If the MaxScale process was started with inadequate permissions to write
into the log file, the mxb_log_finish() would end up being called a
second time. This happened as it was registered to be an atexit function
which ended up getting called when the log initialization failed for the
second time.