MXS-4576: Improve QueryClassifier::update_route_info()
The current_target_t argument was redundant. The only case where it was
something other than CURRENT_TARGET_UNDEFINED was when readwritesplit
was locked to the master. This means that the value of
Handler::is_locked_to_master() can be used to deduce the value to use.

The return value also does not need to be a value. A const reference is
better as it avoids the possibly unnecessary copying of the prepared
statement IDs which can end up being relatively large.

The function currently takes a reference to a GWBUF instead of a const
reference. This was done to avoid having to rewrite all the functions in
one go to support const GWBUFs.