MXS-4544 Add `isAttrToAttr` field to the config of Link
This attribute controls the drawing mode, either from the source node
center to target node center or from source node attribute to target
node attribute. The former mode is the standard mode that other ERD tools
use as it helps to avoid overlapped links while the latter mode is used
by modern ERD tools. Both have pros and cons, so it's better to let the
users change the drawing mode.

This also renames `cols` field in node data to `attrs` because`attributes`
is the name used in ERD.