MXS-4544 Improve dragging events
Dragging node is stopped unexpectedly if the mouse cursor moves faster
than the node. This happens because the `mousemove` event was attached
to a node instead of the HTML document.

This attaches `mousemove` event to document and adds `dragging-mask`
element in `MxsErd` component to prevent `mouseenter` and `mouseleave`
events in other nodes from being triggered.