Add timeouts to PgSQL
The library itself doesn't implement read or write timeouts so we'll
have to emulate them by using the asynchronous API. The waiting is
currently done using a sleep of 10ms but a more efficient method of
implemented using poll() or select() would be better.

Removed the forward declarations of the libpq internal structures and
replaced them with the actual type names. Including the headers should
not cause any problems and if it does, using the pImpl idiom instead of
forward declaring private types would be a cleaner approach.