MXS-4577 Add configuration setting for Postgres parser
With a configuration like, the parser to be used can be specified.

[MyListener]
type=listener
protocol=postgresprotocol
postgresprotocol.parser=[mariadb|pp_postgres|...]

Currently the default is 'mariadb', which means that whatever the
MariaDB protocol is using, will be used. If it is something else,
an attempt to load a module with that name will be made.

Note that this means that different listeners may be using
different parsers. Convenient for testing.