MXS-4588: Collect sescmd history in postgresprotocol
The tracking of session commands is implemented the same way they are in
the mariadbprotocol: all queries that are classified to be routed to all
nodes are stored in the history.

Similar to how the COM_STMT_CLOSE command in the MariaDB protocol
requires the removal of commands from history, the Close command in the
Postgres protocol must also remove the Parse command that defines
it. This is not yet implemented by the current commit as it only covers
storage of Query commands.

In order to know which command is being handled, the client protocol
must track how many queries are being executed and what to do with their
responses. Unlike in the mariadbprotocol module, where this information
is implied by the current state, the postgresprotocol module stores this
information explicitly. This does impose a slight memory overhead that
is less efficient when a very large amount of requests is in flight but
in practice, when there's only a single request being processed, the
overhead should be negligible. It also enables request pipelining and
thus it could be considered a performance improvement in some cases.

Using a std::variant avoids the need to do a memory allocation that
would otherwise be required with an abstract class. If we assume that
most of the commands won't be recorded in the history, this ends up with
less overall memory allocations compared to an abstract base class which
each request type implements.