MXS-4588: Fix Parse ID tracking
As the Parse command does not generate output on its own, the ID from
the GWBUF would be overwritten when the next packet arrived. To
correctly handle a Parse followed by a Sync or a Parse-and-Bind cycle,
the tracked ID should only be set if the value is not zero. This is a
safe comparison as the ID 0 is reserved and not set by the protocol.