Allow N-ary splits in mxb::split
By converting the function into a template function that takes the
number of times to split as its argument, a single function call can be
used to replace repeated mxb::split calls. This can also be done at
compile time which is what the test cases that were added are using.