Place ThisUnit structs in anon namespaces
If more than one ThisUnit struct is defined in the MaxScale core, the
symbols get mixed up and there's a buffer-overflow whenever the ThisUnit
structs are being initialized. The cause of this most likely was the
ThisUnit struct in gateway.cc that got initialized before the ones in
maxscale-common got initialized.