Implement WatchdogNotifier::Workaround with atomics
The use of a secondary thread was not needed as it unconditionally
marked the thread to be working. The theoretical case where the
workaround thread itself would get stuck is also eliminated using this
approach.

Overall, this is expected to be a very minor improvement in memory usage
and the most significant improvement is in the reduction in the size of
stacktraces.