Expose largest session ID ever created
This'll be useful in identifying whether a uint32_t value can be used to
represent all existing sessions in MaxScale. If the maximum ID is larger
than what can be expressed with a uint32_t, other methods of identifying
sessions must be used.

Also made the session_get_next_id() function an internal implementation
detail of the Session class. There's no longer a need to increment it
outside of the session itself since the session is now always created
before any DCBs are created.