MXS-3735 Restructure the layout of ERD
The DDL editor for altering table will be later adjusted for
reusing in ERD. So any new features come in to the DDL editor
will be also available in ERD.

The layout of ERD worksheet is now changed to have three main areas:

* The top toolbar: Show icon buttons to create a new table, generate
ERD from an existing database, and change visualizing style ( e.g. type
of links, relationship type, colors, and size….)

* The graph editor area: Show the diagram.

* The bottom area is for the DDL editor: Double-click on the entity in
the diagram will show this editor.

The DDL editor will be later refactored to either output ALTER TABLE
script or CREATE TABLE script depending on the editor mode.