MXS-3735 Add components to select table objects for generating ERD
This adds `ObjSelectDlg` and `SelectableSchemaTableTree` components
to show schema table tree in a dialog.

Clicking on the "Generate ERD" icon button which is next to the "Reload"
schemas button will show that dialog. Alternative, to generate ERD
for a specific schema, right-clicking on the schema name and
choose "Generate ERD" option.

The `SelectableSchemaTableTree` component can be later reused
in ERD worksheet as there will be a button to generate ERD from
the existing server.