MXS-4600 Do not add '=' as function when it is an assignment
'=' was explicitly injected as a function when the statement
was something like

    INSERT INTO t SET f = 42;

That is non-sensical as the '=' is not a comparison function but
assignment. Further, a side-effect of that was that a statement
like
    INSERT INSERT t (f) VALUES (42);

also appeared to use the '=' function, which then caused a warning
in debug-mode, since '=' obviously was not found in the canonical
statement.

Now '=' is no longer reported to be used and thus the suppression
in complain_about_missing() can be removed. At the same time,
suppression is added for other contructs that are known to
disappear during parsing/canonicalization.