MXS-4551: Don't use compile options with pcre2_match
The options for the mxb::Regex class are intended for the pcre2_compile
call and should not be passed to pcre2_match or pcre2_substitute.

The reason why `options=ignorecase` worked but `options=extended` did
not was because PCRE2_CASELESS had the same value as the
PCRE2_NOTEMPTY_ATSTART option whereas PCRE2_EXTENDED had the same value
as PCRE2_DFA_SHORTEST which was not a valid option for pcre2_match.