MXS-4602: Add test case
Added a test case that reproduces both MXS-4551 and MXS-4602.

In addition, it revealed a problem where multiple ongoing queries end up
causing a loss of information when the newer queries overwrote the SQL
in `m_sql`. This was filed as MXS-4604 but was not included in this test
case as fixing it requires several changes in the qlafilter.