MXS-4601 Fix `escapeIdentifiers` function
If the string contains multiple dots but only have two
identifier names. The function failed to split and quote
identifier names. In addition, if the identifier name contains
backticks, it also failed to quote.

e.g. if the string is: "db.name.table_name". It failes to
recognize db.name as the database name.

This renames the function to `quotingIdentifier` and refactor
it to handle one identifer name. It quotes the identifer name and
escapes the backticks if necessary.
.e.g.
```
test.my test`s ` table -> `test`.`my test``s `` table`
```

When the node is generated for the schema tree, the `qualified_name`
attribute will be quoted. Therefore, the "Qualified Name (Quoted) context
option is now removed as it is no longer needed.

The `getSchemaName` and `getTblName` relies on the dot as the delimiter
to get the corresponding name. e.g. db_name.table_name. However,
the identifier name can contain dots, so using the dot was a bad idea.
This adds a new attribute called `parentNameData` for each node which
contains the name of its ancestor nodes.