MXS-3735 Handle preselected schemas
If there are preselected schemas, it should first load tables of those
schemas and select table nodes.

`mxs-treeview` component is a wrapper component of vuetify `v-treeview`,
so it inherits the `selection-type` props which has `leaf` as the
default mode. If all leaf nodes are all selected, the parent node
won't be included. However, in case the preselect schema has no
table, that schema node should be selected, so the validation input
can be shown to the user.