MXS-3735 Generate ERD data from parsed DDL
This adds functions to generate ERD data from parsed DDL and store
data to a new ERD worksheet.

This adjusts graph data properties so it can used parsed DDL data.

This also renames `ObjSelectDlg` to `GenErdDlg`. This dialog is
specifically used for QueryEditor worksheet. So there is no need
to have a generic name.