Fix alter_router
If the response was an error, the test would crash as it assumed a
non-empty result. Using the field() function makes the comparison safer
as it returns an empty string.

Added an extra check that makes sure the query was
successful. Previously this was implicit and if it failed, no error was
printed.