MXS-4588: Simplify session command history
In some cases, the session command history can be simplified by removing
repeated occurrences of the same command. This works very well for
statements like `SET autocommit=0` and `SET autocommit=1` that get
executed in a cyclical manner. This assumes that the session commands
are idempotent and that they only need to be executed once even if they
are seen multiple times.

This change is expected to reduce the amount of memory used by clients
connected to MaxScale that use client-side connection pools. This also
improves the ability to restore session state for workloads that start
off with a handful of preparatory commands and then proceed to have a
very small cyclical pattern of session commands. Simplifying the history
also speeds up the execution of the commands as there will be less
commands that need to be sent.

The history simplification must preserve the order in which the commands
are executed. This is achieved by always appending new commands at the
end of the history and then removing any duplicates that are found.

The current approach is to do the simplification whenever new commands
are being added. This has the benefit of not losing the session state
information but it does introduce extra work that might not be
needed. However, this does reduce the amount of memory used by MaxScale
even if no history replay ever takes place. An alternative approach
would be to do the simplification when the history is sent which would
require no changes to the existing code.