Remove unnecessary compiler flags
The stack protector is a poor alternative for ASAN and UBSAN which have
proven to be far more effective and detecting errors. The conditional
use of warnings for old compilers is not needed since all supported
platforms use a newer one.

As a result of this, the DEBUG_FLAGS variable can now be removed as it
only contained -ggdb which was also set elsewhere.