MXS-4603 Generate initial servers in monitor thread
The active servers of a monitor should only be modified and read
from the monitor thread as this thread also reacts to server
state changes.

Move the sqlite3-access from post_configure() in the
main thread to pre_run() in the monitor thread. The sqlite-
handle is still opened in the main thread so it should use
internal mutex.