MXS-3735 Improve highlight effect on links
Storing link styles in a temporary property is not a good approach as
mouse events may be delayed, it doesn't guarantee the value in `linkStylesTmp`
won't be overridden by event styles.

The event styles are now stored in `evtLinkStyles` property. After the mouse
event ends, this property will be deleted from the link.

This introduces the `invisibleHighlightColor` property to the link
configuration. As a result, when hovering over or dragging a node,
the previously hidden links become visible with a highlight color,
while the regular links turn white.

This removes the function to get highlight color in `mxs-erd` component
as it's faster to get the value when generating the links.