MXS-3735 Clean up `graph-nodes` and `mxs-erd` borders style
Rename some props in `graph-nodes` and remove `dynHeight` props as
there is no practical usage for that props. The `dynWidth` is renamed to
`autoWidth` as it indicates the behavior of the props clearer.

Replace the methods to get border styles with css code by
inheriting `border-color` set on table element.

This also adds `tdMaxWidth` props in `mxs-erd` to ensure the node
won't exceed 320px in width. The text will be truncated and shown
in tooltip by using `mxs-truncate-str` component.