MXS-4610 Enable tracking the of source of paths
Paths can be set both in the configuration file and provided
on the command line, and the latter should take presedence.

Currently heuristics (is the current value different from
the default) is used for detecting whether a path has been
provided on the command-line, which is in-exact and leads to
the wrong decision if there is a config value and the
command-line provided argument is the same as the default
value. Better to track it explicitly.

The seemingly unrelated change in pp_mysqlembedded.cc is
due to a compilation failure caused by the change in includes.

gateway.cc will be modified in another commit.