MXS-3735 Handle STRAIGHT shape
`STRAIGHT` shape needs a special adjustment for the link and the markers.
Because the angle between source and target point is dynamic, rotating
markers is not sensible as those should be fully visible.

This adds an offset value equals to the width of the marker. So that
link is drawn starting from the edge of the marker.

This also adjusts cardinality symbols to have a consistent width and
removes redundant symbols.